style(Submission): rfc reorg still needs eslint
