+++ src/actions.js
@@ -203,4 +203,14 @@ export const fetchOAuthURL = oauthPath => dispatch =>
         };
       }
     },
-  ).then(msg => dispatch(msg));
+  ).then(
+    (msg) => {
+      dispatch(msg);
+      if ( msg.url ) {
+        console.log( "FETCH OAUTH gOT: " + msg.url );
+        return msg.url;
+      } else {
+        throw new Error('OAuth authorization failed');
+      }
+    }
+  );
