+++ src/index.js
@@ -15,11 +15,10 @@ import ExplorerPage from './Explorer/ExplorerPage';
 import QueryNode, { clearResultAndQuery } from './QueryNode/ReduxQueryNode';
 import DataDictionary from './DataDictionary/ReduxDataDictionary';
 import DataDictionaryNode from './DataDictionary/ReduxDataDictionaryNode';
-import ProjectSubmission from './Submission/component';
+import ProjectSubmission, { loginSubmissionAPI } from './Submission/ReduxProjectSubmission';
 import UserProfile, { loginUserProfile, fetchAccess } from './UserProfile/ReduxUserProfile';
 import CertificateQuiz from './Certificate/ReduxQuiz';
 import GraphQLQuery, { fetchSchema } from './GraphQLEditor/ReduxGqlEditor';
-import { loginSubmissionAPI } from './Submission/actions';
 import { fetchDictionary } from './queryactions';
 import { app } from './localconf';
 import browserHistory from './history';
